Beef up check_data. Now collects stats on columns in use in the CSV files:
Each time a column is in use it is counted, and the stats for each col are output.
Lets you see fairly easily which cols are necessary to expose to the user.
Suspect there is a fair amount of overlap between this and the cols harvested from
the historical code. Will have to handle all cases though of course:
getting cols from one CSV that aren't in the next, etc. etc.